# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenLane API
  version: 0.1.0
  description: >
    The OpenLane public API — source of truth (ADR-0005).
    Every UI action hits this API. Versioned under /v1; breaking changes need /v2.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.txt
servers:
  - url: http://localhost:8080/v1
    description: local
paths:
  /portal-links:
    post:
      summary: Create a portal link for a project contact
      description: >
        Internal (staff) only. Generates a 32-byte random token; the DB stores
        only its SHA-256 hash. Token is returned exactly once at creation.
      operationId: createPortalLink
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, contact_id]
              properties:
                project_id:
                  type: string
                  format: uuid
                contact_id:
                  type: string
                  format: uuid
                expires_in_days:
                  type: integer
                  minimum: 1
                  maximum: 30
                  default: 7
      responses:
        "201":
          description: Portal link created; token shown exactly once
          content:
            application/json:
              schema:
                type: object
                required: [id, project_id, contact_id, token, expires_at]
                properties:
                  id:
                    type: string
                    format: uuid
                  project_id:
                    type: string
                    format: uuid
                  contact_id:
                    type: string
                    format: uuid
                  token:
                    type: string
                    description: Raw token — shown exactly once, never stored
                  expires_at:
                    type: string
                    format: date-time
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      summary: List portal links for a project
      operationId: listPortalLinks
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Portal links (never includes raw tokens)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortalLink"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /portal-links/{id}/revoke:
    post:
      summary: Revoke a portal link
      description: Revocation is immediate; the token stops working on the next request.
      operationId: revokePortalLink
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /portal/{token}/session:
    get:
      summary: Resolve a portal session
      description: >
        Validates the portal token, returns workspace/project/customer context
        and the contact it belongs to. RLS scope for everything else flows from here.
      operationId: getPortalSession
      security:
        - portalToken: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 43
            maxLength: 128
      responses:
        "200":
          description: Active session for this token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "410":
          $ref: "#/components/responses/Gone"
  /portal/{token}/tasks:
    get:
      summary: List customer-visible tasks for the portal session
      description: >
        RLS already restricts to the token's project and customer_visible=true
        rows; the handler is a plain SELECT. EC-PORTAL-07 defense lives in the DB.
      operationId: listPortalTasks
      security:
        - portalToken: []
      parameters:
        - $ref: "#/components/parameters/PortalTokenPathParam"
      responses:
        "200":
          description: Customer-visible tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortalTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "410":
          $ref: "#/components/responses/Gone"
  /portal/{token}/tasks/{task_id}/complete:
    post:
      summary: Mark a customer-visible task complete
      description: >
        Sets status=done + completed_at. RLS permits only customer_visible tasks
        in the token's project; internal tasks return 404 (not 403 — no oracle).
      operationId: completePortalTask
      security:
        - portalToken: []
      parameters:
        - $ref: "#/components/parameters/PortalTokenPathParam"
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
...
  /v1/imports:
    post:
      summary: Import tasks from a competitor CSV (dry-run or commit)
      description: |
        Parses an Asana/Rocketlane/generic CSV export into OpenLane tasks.
        With dry_run=true nothing is written; the response is the full
        per-row validation report. Without it, valid rows are inserted in
        one transaction and invalid rows are quarantined in the response.
      operationId: createImport
      security:
        - bearerAuth: []
      parameters:
        - name: dry_run
          in: query
          schema:
            type: boolean
            default: false
        - name: X-Workspace-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, source]
              properties:
                file:
                  type: string
                  format: binary
                source:
                  type: string
                  enum: [asana, rocketlane, generic]
                project_id:
                  type: string
                  format: uuid
                  description: Existing project to import into (required unless create_project given).
                create_project:
                  type: object
                  properties:
                    customer_id:
                      type: string
                      format: uuid
                    name:
                      type: string
                      minLength: 3
                      maxLength: 120
                  required: [customer_id, name]
      responses:
        "200":
          description: Validation report (dry run) or import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Bad request (unknown source, missing file, no project target)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Target project not found (or out of workspace scope)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /projects:
    post:
      summary: Create a project
      operationId: createProject
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer_id, name]
              properties:
                customer_id: {type: string, format: uuid}
                name: {type: string, minLength: 3, maxLength: 120}
                status:
                  type: string
                  enum: [draft, active]
                  default: draft
                start_date: {type: string, format: date}
                target_go_live: {type: string, format: date}
                template_id: {type: string, format: uuid, description: Optional — create tasks from template phases}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Project"}
        "400":
          description: Validation (name length, start > target, unknown status)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
        "404":
          description: Customer not found (or out of scope)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
    get:
      summary: List projects (filters)
      operationId: listProjects
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: status
          in: query
          schema: {type: string, enum: [draft, active, on_hold, at_risk, delayed, completed, cancelled]}
        - name: customer_id
          in: query
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Projects in scope
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/Project"}
        "401":
          description: Invalid staff credentials
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /projects/{id}:
    get:
      summary: Project detail (with computed progress)
      operationId: getProject
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Project"}
        "404":
          description: Not found (or out of workspace scope)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
    patch:
      summary: Update project (status transitions guarded per spec §6.3)
      operationId: patchProject
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string, minLength: 3, maxLength: 120}
                status: {type: string, enum: [draft, active, on_hold, at_risk, delayed, completed, cancelled]}
                start_date: {type: string, format: date}
                target_go_live: {type: string, format: date}
                health: {type: string, enum: [green, amber, red, unknown]}
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Project"}
        "400":
          description: Transition guard failed (e.g. draft→active without start_date; active→completed with open required tasks)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
        "404":
          description: Not found
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
    delete:
      summary: Soft-delete (archive) a project
      operationId: deleteProject
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "204": {description: Archived}
        "404":
          description: Not found
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /templates:
    post:
      summary: Create a template (v1.0.0)
      operationId: createTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TemplateInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Template"}
        "400":
          description: Validation (phases/tasks shape)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
        "401":
          description: Invalid staff credentials
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
    get:
      summary: List templates
      operationId: listTemplates
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: category
          in: query
          schema: {type: string}
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/Template"}
        "401":
          description: Invalid staff credentials
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /templates/{id}:
    get:
      summary: Template detail
      operationId: getTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Template"}
        "404":
          description: Not found
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
...
  /projects/{id}/tasks:
    get:
      summary: List project tasks (staff — includes internal)
      operationId: listProjectTasks
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: status
          in: query
          schema: {type: string}
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/StaffTask"}
        "404":
          description: Project not found (or out of scope)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
    post:
      summary: Create a task in the project
      operationId: createTask
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TaskInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StaffTask"}
        "400":
          description: Validation (title length, invalid status)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
        "404":
          description: Project not found
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /tasks/{id}:
    patch:
      summary: Update task (§8.1 state machine enforced)
      operationId: patchTask
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/TaskInput"
                - type: object
                  properties:
                    reopen_reason: {type: string, maxLength: 500, description: Required when status moves done→todo}
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StaffTask"}
        "400":
          description: Transition guard failed (e.g. done→todo without reason)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
        "404":
          description: Task not found
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
    delete:
      summary: Soft-delete a task
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - name: X-Workspace-Id
          in: header
          required: true
          schema: {type: string, format: uuid}
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "204": {description: Deleted}
        "404":
          description: Not found
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /auth/magic-link:
    post:
      summary: Request a magic-link login token (email + workspace slug)
      description: >
        Always returns 202 (never reveal whether the email exists). In dev
        mode (OPENLANE_DEV_LOGIN=1) the response includes dev_token so
        flows work without SMTP. Token is single-use, 15 minutes.
      operationId: requestMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, workspace_slug]
              properties:
                email: {type: string, format: email}
                workspace_slug: {type: string}
      responses:
        "202":
          description: Token sent (dev_token included when dev login enabled)
          content:
            application/json:
              schema:
                type: object
                properties:
                  dev_token: {type: string}
        "400":
          description: Malformed request
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /auth/magic-link/consume:
    get:
      summary: Exchange a magic-link token for session tokens
      operationId: consumeMagicLink
      parameters:
        - name: token
          in: query
          required: true
          schema: {type: string, minLength: 43}
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionTokens"
        "401":
          description: Invalid, expired, or already-used token
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /auth/refresh:
    post:
      summary: Rotate a refresh token (family theft detection)
      description: >
        Refresh tokens rotate on every use. Presenting an already-rotated
        token revokes the entire family (assumed theft).
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string, minLength: 43}
      responses:
        "200":
          description: New session tokens
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionTokens"}
        "401":
          description: Unknown or revoked token (family revoked if reuse)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /auth/logout:
    post:
      summary: Revoke the refresh family (bearer auth)
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string, minLength: 43}
      responses:
        "204": {description: Logged out}
        "401":
          description: Bad token
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /auth/me:
    get:
      summary: Current user + workspace + role (bearer auth)
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Membership context
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Me"}
        "401":
          description: Invalid or expired JWT
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /auth/workspaces:
    get:
      summary: List workspaces the user belongs to (bearer auth)
      operationId: listMyWorkspaces
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Memberships
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/Me"}
        "401":
          description: Invalid JWT
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PortalTokenPathParam:
      name: token
      in: path
      required: true
      schema:
        type: string
        minLength: 43
        maxLength: 128
  responses:
    BadRequest:
      description: Malformed request (RFC 9457 problem+json)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing, invalid, expired, or revoked credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource does not exist or is outside the caller's scope (no oracle)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Gone:
      description: Portal link revoked or expired
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ImportResult:
      type: object
      properties:
        dry_run:
          type: boolean
        source:
          type: string
        project_id:
          type:
            - string
            - "null"
          format: uuid
        ok_count:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              source_row:
                type: integer
              reason:
                type: string
...
    PortalLink:
      type: object
      required: [id, project_id, contact_id, status, expires_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        contact_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, revoked, expired]
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type:
            - string
            - "null"
          format: date-time
        created_at:
          type: string
          format: date-time
    PortalSession:
      type: object
      required: [project_id, project_name, customer_name, contact_id, contact_name, expires_at]
      properties:
        project_id:
          type: string
          format: uuid
        project_name:
          type: string
        customer_name:
          type: string
        contact_id:
          type: string
          format: uuid
        contact_name:
          type: string
        expires_at:
          type: string
          format: date-time
    PortalTask:
      type: object
      required: [id, project_id, title, status, customer_visible, due_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [todo, in_progress, blocked, review, done, waived]
        customer_visible:
          type: boolean
        due_at:
          type:
            - string
            - "null"
          format: date-time
        completed_at:
          type:
            - string
            - "null"
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    portalToken:
      type: apiKey
      in: query
      name: portal_token
security:
  - bearerAuth: []